appEngine/requirements.txt
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
mysql-connector-python==8.0.23
SQLAlchemy==1.3.8
google-api-core==1.26.0
google-api-python-client==2.0.2
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-cloud-core==1.6.0
google-cloud-storage==1.36.0
google-cloud-pubsub==2.3.0
google-cloud-logging==2.3.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
oauth2client==4.1.3
